refactor: 라인 삭제 시, 구간을 삭제하고 검증하는 로직 추가

+++ src/main/java/wooteco/subway/service/LineService.java
@@ -111,7 +111,9 @@ public class LineService {
 
     @Transactional
     public void deleteLine(final long id) {
-        sectionDao.deleteByLineId(id);
+        if (sectionDao.deleteByLineId(id) == NOT_FOUND) {
+            throw new StationNotFoundException();
+        }
         if (lineDao.delete(id) == NOT_FOUND) {
             throw new LineNotFoundException();
         }
